Moved ThreadSafeMonoCecilTempGenerator to helpers directory
